+++ tests/testthat/test_dirichlet_hmm.R
@@ -3,6 +3,8 @@ context("Hidden Markov Model")
 testData <- c(rnorm(50, 1, sqrt(3)), rnorm(50, 3, sqrt(3)), rnorm(50, 5, sqrt(3)))
 normMD <- GaussianMixtureCreate()
 
+expNames <- c("data", "n", "mixingDistribution", "states", "uniqueParams", "params", "alpha", "beta")
+
 HMM_dp_test <- function(dp){
 
   expect_is(dp, c("list", "markov", "dirichetprocess", "normal", "conjugate"))
@@ -18,13 +20,10 @@ test_that("Create",{
 
   dp <- DirichletHMMCreate(testData, normMD, 2, 3)
 
-
-  expectedValues <- c("data", "mdobj", "states", "params", "alpha", "beta")
+  expectedValues <- expNames
   expect_equal(names(dp), expectedValues)
 
   HMM_dp_test(dp)
-
-
 })
 
 test_that("Update States Integration", {
@@ -33,11 +32,10 @@ test_that("Update States Integration", {
 
   dp <- UpdateStates(dp)
 
-  expectedValues <- c("data", "mdobj", "states", "params", "alpha", "beta")
+  expectedValues <- expNames
   expect_equal(names(dp), expectedValues)
 
   HMM_dp_test(dp)
-
 })
 
 test_that("Update Parameters Integration", {
@@ -46,7 +44,7 @@ test_that("Update Parameters Integration", {
   dp <- UpdateStates(dp)
   dp <- param_update(dp)
 
-  expectedValues <- c("data", "mdobj", "states", "params", "alpha", "beta")
+  expectedValues <- expNames
   expect_equal(names(dp), expectedValues)
 
 })